refactor: extract TriadAmsInputMediaPlayer into separate module
- Create input_media_player.py (453 lines) containing:
  - TriadAmsInputMediaPlayer class
  - InvalidGroupMemberError exception
- Update media_player.py (780 lines, down from 1,197)
  - Import TriadAmsInputMediaPlayer from new module
  - Remove unused asyncio import
- Improves code organization and maintainability
- All 282 tests pass (264 unit + 18 integration)
